requirements.txt
# Core runtime dependencies for AQ-NEW-NL2SQL
# Notes:
# - langchain-openai provides AzureChatOpenAI bindings used in nl2sql_main.py
# - langchain for ChatPromptTemplate and chains
# - python-dotenv for loading environment variables from .env
# - pyodbc for Azure SQL Database connectivity (requires Microsoft ODBC Driver 18 installed on the OS)

langchain>=0.2.12,<0.3
langchain-openai>=0.1.8,<0.2
python-dotenv>=1.0.1,<2.0.0
pyodbc>=5.1.0,<6.0.0
requests>=2.32.0,<3.0.0
streamlit>=1.36,<2.0
pandas>=2.1,<3.0
XlsxWriter>=3.1,<4.0
watchdog>=4.0,<5.0
azure-storage-blob>=12.0.0
PyYAML>=6.0,<7.0  # for data-driven NL intent catalog
